move tests to right file...
