+++ tests/Integration/Execution/MutationExecutor/MorphToManyTest.php
@@ -2,11 +2,8 @@
 
 namespace Tests\Integration\Execution\MutationExecutor;
 
-use Faker\Provider\Lorem;
 use Tests\DBTestCase;
-use Tests\Utils\Models\Role;
 use Tests\Utils\Models\Tag;
-use Tests\Utils\Models\User;
 
 class MorphToManyTest extends DBTestCase
 {
@@ -14,7 +11,6 @@ class MorphToManyTest extends DBTestCase
     type Mutation {
         createTask(input: CreateTaskInput! @spread): Task @create
         upsertTask(input: UpsertTaskInput! @spread): Task @upsert
-        updateUser(input: UpdateUserInput! @spread): User @update
     }
 
     input CreateTaskInput {
@@ -61,37 +57,6 @@ class MorphToManyTest extends DBTestCase
         id: ID!
         name: String!
     }
-
-    input UpdateUserInput {
-        id: ID!
-        roles: UpdateRoleRelation
-    }
-
-    input UpdateRoleRelation {
-        sync: [UpdateUserRolePivot!]
-        syncWithoutDetaching: [UpdateUserRolePivot!]
-        connect: [UpdateUserRolePivot!]
-    }
-
-    input UpdateUserRolePivot {
-        id: ID! # role ID
-        meta: String
-    }
-
-    type User {
-        id: ID!
-        roles: [Role!]!
-    }
-
-    type Role {
-        id: ID!
-        pivot: UserRolePivot
-    }
-
-    type UserRolePivot {
-        meta: String
-    }
-
     '.self::PLACEHOLDER_QUERY;
 
     public function testCanCreateATaskWithExistingTagsByUsingConnect(): void
@@ -178,141 +143,6 @@ class MorphToManyTest extends DBTestCase
                     'tags' => [
                         [
                             'id' => $id,
-                        ],
-                    ],
-                ],
-            ],
-        ]);
-    }
-
-    public function testCanConnectUserWithRoleAndPivotMetaByUsingSync(): void
-    {
-        $user = factory(User::class)->create();
-        $role = factory(Role::class)->create();
-        $role2 = factory(Role::class)->create();
-        $user->roles()->attach($role2);
-
-        $metaText = Lorem::sentence();
-
-        $this->graphQL('
-        mutation {
-            updateUser(input: {
-                id: 1,
-                roles: {
-                    sync: [
-                        {
-                            id: 1,
-                            meta: "'.$metaText.'"
-                        }
-                    ]
-                },
-            }) {
-                roles {
-                    pivot {
-                        meta
-                    }
-                }
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'updateUser' => [
-                    'roles' => [
-                        [
-                            'pivot' => [
-                                'meta' => $metaText,
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ]);
-    }
-
-    public function testCanConnectUserWithRoleAndPivotMetaByUsingSyncWithoutDetach(): void
-    {
-        $user = factory(User::class)->create();
-        $role = factory(Role::class)->create();
-        $role2 = factory(Role::class)->create();
-        $user->roles()->attach($role2);
-
-        $metaText = Lorem::sentence();
-
-        $this->graphQL('
-        mutation {
-            updateUser(input: {
-                id: 1,
-                roles: {
-                    syncWithoutDetaching: [
-                        {
-                            id: 1,
-                            meta: "'.$metaText.'"
-                        }
-                    ]
-                },
-            }) {
-                roles {
-                    pivot {
-                        meta
-                    }
-                }
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'updateUser' => [
-                    'roles' => [
-                        [
-                            'pivot' => [
-                                'meta' => $metaText,
-                            ],
-                        ],
-                        [
-                            'pivot' => [
-                                'meta' => null,
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ]);
-    }
-
-    public function testCanConnectUserWithRoleAndPivotMetaByUsingConnect(): void
-    {
-        $user = factory(User::class)->create();
-        $role = factory(Role::class)->create();
-
-        $metaText = Lorem::sentence();
-
-        $this->graphQL('
-        mutation {
-            updateUser(input: {
-                id: 1,
-                roles: {
-                    connect: [
-                        {
-                            id: 1,
-                            meta: "'.$metaText.'"
-                        }
-                    ]
-                },
-            }) {
-                roles {
-                    pivot {
-                        meta
-                    }
-                }
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'updateUser' => [
-                    'roles' => [
-                        [
-                            'pivot' => [
-                                'meta' => $metaText,
-                            ],
                         ],
                     ],
                 ],
